Move `AblyObjectsTypes` definition to `index-objects` file in package test
